move starpak structs to starpak.h
